Reorganize crypto backends
