+++ docs/reference/CryptoBackends/README.md
@@ -2,7 +2,7 @@
 
 This contains the code to integrate libgossamer with various cryptography
 backends that provide cryptographic signatures. This allows projects to
-extend the [Signer](../Signer.md) and [Verifier](../Verifier.md) classes
+extend the [Signer](../Release/Signer.md) and [Verifier](../Release/Verifier.md) classes
 to support Hardware Security Modules (HSMs).
 
 We ship with a single backend (`SodiumBackend`) which uses Libsodium to
