+++ docs/reference/Release/Common.md
@@ -14,8 +14,8 @@
 **Arguments**:
 
 1. `int` Algorithm (see [Class Constants](#class-constants))
-2. [`CryptoBackendInterface`](CryptoBackendInterface.md) Optional backend.
-   Defaults to [SodiumBackend](Backends/SodiumBackend.md).
+2. [`CryptoBackendInterface`](../CryptoBackendInterface.md) Optional backend.
+   Defaults to [SodiumBackend](../CryptoBackends/SodiumBackend.md).
 
 ## Static Methods
