+++ docs/reference/Release/README.md
@@ -5,9 +5,9 @@ version releases for packages.
 
 ## Classes
 
-* [Backends](Backends)
-    * [SodiumBackend](Backends/SodiumBackend.md)
+* [Backends](../CryptoBackends)
+    * [SodiumBackend](../CryptoBackends/SodiumBackend.md)
 * [Common](Common.md)
-* [CryptoBakendInterface](CryptoBackendInterface.md)
+* [CryptoBakendInterface](../CryptoBackendInterface.md)
 * [Signer](Signer.md)
 * [Verifier](Verifier.md)
